Fix a bug in the load promise of embedded collections
The _meta of an embedded collection should always mirror the _meta of
the containing entity. While reloading an embedded collection, the
loading flag is kept falsey on both containing entity and embedded
collection. However, the load promise was changed on the containing
entity while the load promise of the embedded collection kept on
directly resolving to the pre-reload data during the reload.

This commit fixes this by basing the embedded collection load promise on
the load promise of the containing entity, during (re-)loads.